feat
se crearon nuevas vistas, se modularon otras ya existentes como la prevista de pago y la vista de pago
