+++ client/marketplace-frontend/src/componentes/pagePay/pagePay.js
@@ -1,0 +1,219 @@
+import React, { useState, useEffect } from 'react';
+import './pagePay.css';
+import { useNavigate } from 'react-router-dom';
+import { useCart } from '../../context/CartContext';
+
+const PagePay = () => {
+    const navigate = useNavigate();
+    const { cartItems, cartTotal, removeFromCart, updateQuantity, clearCart } = useCart();
+
+    // Estado para items seleccionados
+    const [selectedItems, setSelectedItems] = useState({});
+
+    // Inicializar selección (todos seleccionados al cargar)
+    useEffect(() => {
+        const initial = {};
+        cartItems.forEach(item => {
+            initial[item.id] = true;
+        });
+        setSelectedItems(initial);
+    }, [cartItems.length]);
+
+    const [isProcessing, setIsProcessing] = useState(false);
+    const [paymentSuccess, setPaymentSuccess] = useState(false);
+
+    // Cálculos
+    const selectedCount = cartItems.filter(item => selectedItems[item.id]).length;
+    const isAllSelected = cartItems.length > 0 && selectedCount === cartItems.length;
+
+    const selectedSubtotal = cartItems.reduce((acc, item) => {
+        return selectedItems[item.id] ? acc + (item.price * item.quantity) : acc;
+    }, 0);
+
+    const toggleSelection = (id) => {
+        setSelectedItems(prev => ({
+            ...prev,
+            [id]: !prev[id]
+        }));
+    };
+
+    const toggleSelectAll = () => {
+        if (isAllSelected) {
+            setSelectedItems({});
+        } else {
+            const allSelected = {};
+            cartItems.forEach(item => {
+                allSelected[item.id] = true;
+            });
+            setSelectedItems(allSelected);
+        }
+    };
+
+    const handleRemove = (id) => {
+        removeFromCart(id);
+        const newSelected = { ...selectedItems };
+        delete newSelected[id];
+        setSelectedItems(newSelected);
+    };
+
+    const handleSubmit = (e) => {
+        e.preventDefault();
+        if (selectedCount === 0) {
+            alert('Por favor selecciona al menos un producto');
+            return;
+        }
+        setIsProcessing(true);
+        setTimeout(() => {
+            setIsProcessing(false);
+            setPaymentSuccess(true);
+            clearCart();
+        }, 2000);
+    };
+
+    if (paymentSuccess) {
+        return (
+            <div className="pay-container success-view">
+                <div className="success-card">
+                    <div className="success-icon">✓</div>
+                    <h1>¡Pedido Realizado!</h1>
+                    <p>Gracias por tu compra. Te contactaremos pronto.</p>
+                    <button className="back-home-btn" onClick={() => navigate('/')}>Volver a la tienda</button>
+                </div>
+            </div>
+        );
+    }
+
+    if (cartItems.length === 0) {
+        return (
+            <div className="pay-container empty-view">
+                <div className="empty-card">
+                    <h1>Tu carrito está vacío</h1>
+                    <p>Agrega productos para comenzar tu compra.</p>
+                    <button className="back-home-btn" onClick={() => navigate('/')}>Explorar productos</button>
+                </div>
+            </div>
+        );
+    }
+
+    return (
+        <div className="pay-container">
+            {/* Header / Breadcrumbs */}
+            <div className="pay-header-simple">
+                <div className="breadcrumb">
+                    <span onClick={() => navigate('/')}>Inicio</span> &gt; <span>Carrito</span>
+                </div>
+            </div>
+
+            <div className="pay-main-content">
+                {/* Lado Izquierdo: Lista de Productos */}
+                <div className="products-column">
+                    <div className="selection-bar">
+                        <div className="select-all-group" onClick={toggleSelectAll}>
+                            <input
+                                type="checkbox"
+                                checked={isAllSelected}
+                                readOnly
+                            />
+                            <span>Seleccionar todo ({cartItems.length})</span>
+                        </div>
+                        <div className="selection-stats">
+                            <span>Sugeridos ({cartItems.length})</span>
+                            <span className="selected-active">Seleccionado ({selectedCount})</span>
+                        </div>
+                    </div>
+
+                    <div className="checkout-items-list">
+                        {cartItems.map(item => (
+                            <div key={item.id} className={`checkout-item-row ${!selectedItems[item.id] ? 'dimmed' : ''}`}>
+                                <div className="item-checkbox">
+                                    <input
+                                        type="checkbox"
+                                        checked={!!selectedItems[item.id]}
+                                        onChange={() => toggleSelection(item.id)}
+                                    />
+                                </div>
+
+                                <div className="item-main-info">
+                                    <div className="item-image-box">
+                                        <img src={item.images[0]} alt={item.name} />
+                                    </div>
+                                    <div className="item-details-box">
+                                        <h4 className="item-name-checkout">{item.name}</h4>
+                                        <p className="item-vendor-checkout">Nexora Premium</p>
+                                        <div className="item-pricing-row">
+                                            <span className="current-price">${item.price}</span>
+                                        </div>
+                                    </div>
+                                </div>
+
+                                <div className="item-actions-box">
+                                    <button className="delete-item-btn" onClick={() => handleRemove(item.id)}>🗑️</button>
+                                    <div className="quantity-selector-checkout">
+                                        <span>Cant. </span>
+                                        <select
+                                            value={item.quantity}
+                                            onChange={(e) => updateQuantity(item.id, parseInt(e.target.value))}
+                                        >
+                                            {[1, 2, 3, 4, 5, 6, 7, 8, 9, 10].map(n => (
+                                                <option key={n} value={n}>{n}</option>
+                                            ))}
+                                        </select>
+                                    </div>
+                                </div>
+                            </div>
+                        ))}
+                    </div>
+                </div>
+
+                {/* Lado Derecho: Resumen Sticky */}
+                <aside className="summary-sidebar">
+                    <div className="sticky-summary-card">
+                        <h3>Resumen del pedido</h3>
+
+                        <div className="summary-details">
+                            <div className="detail-line">
+                                <span>Total de artículos:</span>
+                                <span>${selectedSubtotal.toFixed(2)}</span>
+                            </div>
+                            <div className="detail-line shipping">
+                                <span>Envío:</span>
+                                <span>GRATIS</span>
+                            </div>
+                        </div>
+
+                        <div className="summary-total-final">
+                            <div className="total-label-row">
+                                <strong>Total</strong>
+                                <span className="total-amount-large">${selectedSubtotal.toFixed(2)}</span>
+                            </div>
+                            <p className="tax-hint">Consulta el monto final al completar el pago.</p>
+                        </div>
+
+                        <button
+                            className="order-submit-btn-temu"
+                            disabled={isProcessing || selectedCount === 0}
+                            onClick={handleSubmit}
+                        >
+                            {isProcessing ? 'Procesando...' : `Hacer pedido (${selectedCount})`}
+                        </button>
+
+                        <div className="payment-trust-section">
+                            <p className="secure-payment-text">🛡️ Opciones de pago seguro</p>
+                            <div className="payment-methods-grid">
+                                <span>VISA</span>
+                                <span>MasterCard</span>
+                                <span>PayPal</span>
+                                <span>Apple Pay</span>
+                            </div>
+                            <p className="trust-disclaimer">
+                                Nexora se compromete a proteger tu información de pago.
+                            </p>
+                        </div>
+                    </div>
+                </aside>
+            </div>
+        </div>
+    );
+};
+
+export default PagePay;
