+++ client/marketplace-frontend/src/componentes/pageViewProduct/pageViewProduct.css
@@ -1,0 +1,334 @@
+/* src/componentes/pageViewProduct/pageViewProduct.css */
+:root {
+    --nexora-blue: #0094ff;
+    --nexora-bg: #0a0e27;
+    --nexora-border: rgba(0, 148, 255, 0.2);
+}
+
+.product-page-container {
+    min-height: 100vh;
+    background-color: var(--nexora-bg);
+    color: white;
+    padding: 40px 20px;
+    font-family: 'Inter', sans-serif;
+}
+
+.loading {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    font-size: 24px;
+    color: var(--nexora-blue);
+}
+
+.product-page-content {
+    max-width: 2000px;
+    margin: 0 auto;
+}
+
+.back-btn {
+    background: transparent;
+    border: 1px solid var(--nexora-border);
+    color: var(--nexora-blue);
+    padding: 10px 20px;
+    border-radius: 8px;
+    cursor: pointer;
+    margin-bottom: 30px;
+    transition: all 0.3s ease;
+}
+
+.back-btn:hover {
+    background: rgba(0, 148, 255, 0.1);
+    transform: translateX(-5px);
+}
+
+.main-content-layout {
+    display: flex;
+    gap: 40px;
+    align-items: flex-start;
+}
+
+.image-section {
+    flex: 1.2;
+    max-width: 1200px;
+}
+
+.main-image-container {
+    position: relative;
+    width: 100%;
+    aspect-ratio: 16 / 10;
+    background: rgba(0, 148, 255, 0.05);
+    border: 1px solid var(--nexora-border);
+    border-radius: 16px;
+    overflow: hidden;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.product-main-image {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+}
+
+.image-nav-btn {
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    background: rgba(0, 0, 0, 0.5);
+    color: white;
+    border: none;
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    cursor: pointer;
+    z-index: 2;
+}
+
+.image-nav-btn.prev {
+    left: 10px;
+}
+
+.image-nav-btn.next {
+    right: 10px;
+}
+
+.image-page-indicator {
+    position: absolute;
+    bottom: 10px;
+    right: 10px;
+    background: rgba(0, 0, 0, 0.6);
+    padding: 4px 12px;
+    border-radius: 20px;
+    font-size: 12px;
+}
+
+.image-thumbnails {
+    display: flex;
+    gap: 10px;
+    margin-top: 20px;
+    overflow-x: auto;
+    padding-bottom: 10px;
+}
+
+.thumbnail {
+    width: 80px;
+    height: 80px;
+    border-radius: 8px;
+    border: 2px solid transparent;
+    cursor: pointer;
+    overflow: hidden;
+    flex-shrink: 0;
+}
+
+.thumbnail.active {
+    border-color: var(--nexora-blue);
+}
+
+.thumbnail img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.info-section {
+    flex: 2;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    position: sticky;
+    top: 20px;
+}
+
+.product-name {
+    font-size: 40px;
+    color: var(--nexora-blue);
+    margin: 0;
+}
+
+.product-meta {
+    display: flex;
+    gap: 20px;
+    align-items: center;
+}
+
+.rating {
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    font-size: 18px;
+}
+
+.vendor {
+    color: #9ca3af;
+}
+
+.price-tag {
+    padding: 20px;
+    background: rgba(16, 185, 129, 0.1);
+    border: 1px solid rgba(16, 185, 129, 0.2);
+    border-radius: 12px;
+    width: fit-content;
+}
+
+.product-price {
+    font-size: 36px;
+    font-weight: 800;
+    color: #10b981;
+}
+
+.description-box h3 {
+    color: var(--nexora-blue);
+    margin-bottom: 10px;
+}
+
+.description-box p {
+    line-height: 1.6;
+    color: #d1d5db;
+}
+
+.action-buttons {
+    margin-top: 10px;
+}
+
+.add-to-cart-btn {
+    width: 100%;
+    max-width: 400px;
+    background: linear-gradient(135deg, #0094ff 0%, #00c2ff 100%);
+    color: white;
+    border: none;
+    padding: 16px;
+    border-radius: 12px;
+    font-size: 18px;
+    font-weight: 700;
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.add-to-cart-btn:hover {
+    transform: translateY(-3px);
+    box-shadow: 0 10px 20px rgba(0, 148, 255, 0.3);
+}
+
+.notification {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+    background: #10b981;
+    color: white;
+    padding: 15px 25px;
+    border-radius: 8px;
+    z-index: 1000;
+    animation: slideIn 0.3s ease;
+}
+
+@keyframes slideIn {
+    from {
+        transform: translateX(100%);
+    }
+
+    to {
+        transform: translateX(0);
+    }
+}
+
+.comments-section-list {
+    margin-top: 50px;
+    padding-top: 30px;
+    border-top: 2px solid var(--nexora-border);
+}
+
+.comments-section-list h3 {
+    color: var(--nexora-blue);
+    margin-bottom: 20px;
+}
+
+.comment-input {
+    width: 100%;
+    background: #1a1f3a;
+    border: 1px solid var(--nexora-border);
+    color: white;
+    padding: 15px;
+    border-radius: 8px;
+    margin-bottom: 15px;
+    min-height: 100px;
+}
+
+.comment-submit-btn {
+    background: #7c3aed;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+.comment-item {
+    background: rgba(255, 255, 255, 0.05);
+    padding: 15px;
+    border-radius: 8px;
+    margin-bottom: 10px;
+}
+
+/* Similar Products */
+.similar-products-section {
+    margin-top: 50px;
+}
+
+.similar-products-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+    gap: 20px;
+    margin-top: 20px;
+}
+
+.similar-product-card {
+    background: #1a1f3a;
+    border: 1px solid var(--nexora-border);
+    border-radius: 12px;
+    overflow: hidden;
+    cursor: pointer;
+    transition: transform 0.3s ease;
+}
+
+.similar-product-card:hover {
+    transform: translateY(-5px);
+}
+
+.similar-image {
+    height: 150px;
+    overflow: hidden;
+}
+
+.similar-image img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.similar-info {
+    padding: 15px;
+}
+
+.similar-name {
+    margin: 0;
+    font-weight: 600;
+}
+
+.similar-price {
+    color: #fbbf24;
+    font-weight: 800;
+    margin: 5px 0;
+}
+
+@media (max-width: 900px) {
+    .main-content-layout {
+        flex-direction: column;
+    }
+
+    .image-section {
+        max-width: 100%;
+    }
+}
