+++ client/marketplace-frontend/src/context/CartContext.js
@@ -22,11 +22,11 @@ export const CartProvider = ({ children }) => {
 
   const addToCart = (product) => {
     setCartItems((prevItems) => {
-      const existingItem = prevItems.find((item) => item._id === product._id);
+      const existingItem = prevItems.find((item) => item.id === product.id);
       if (existingItem) {
         // Si ya existe, aumentar cantidad
         return prevItems.map((item) =>
-          item._id === product._id
+          item.id === product.id
             ? { ...item, quantity: item.quantity + 1 }
             : item,
         );
@@ -39,7 +39,7 @@ export const CartProvider = ({ children }) => {
 
   const removeFromCart = (productId) => {
     setCartItems((prevItems) =>
-      prevItems.filter((item) => item._id !== productId),
+      prevItems.filter((item) => item.id !== productId),
     );
   };
 
@@ -47,7 +47,7 @@ export const CartProvider = ({ children }) => {
     if (newQuantity < 1) return;
     setCartItems((prevItems) =>
       prevItems.map((item) =>
-        item._id === productId ? { ...item, quantity: newQuantity } : item,
+        item.id === productId ? { ...item, quantity: newQuantity } : item,
       ),
     );
   };
